create README.md and fix migration file

+++ src/main/resources/db/migration/V1__create_books_table.sql
@@ -1,7 +1,7 @@
 CREATE TYPE published_status AS ENUM ('UNPUBLISHED', 'PUBLISHED');
 
 CREATE TABLE books (
-    id bigint PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
     price DECIMAL(10, 2) CHECK (price >= 0),
     author VARCHAR(255) NOT NULL,
