+++ src/main/resources/db/migration/V2__create_authors_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE authors (
-    id bigint PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     birth_date DATE NOT NULL CHECK (birth_date < CURRENT_DATE)
 );
